% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{button}
\alias{button}
\title{Button}
\usage{
button(id, ..., class = ...class(), .tag = htmltools::tags$button, value = 0L)
}
\arguments{
\item{id}{ID of button and \code{input}.}

\item{...}{Content of button.}

\item{class}{Any additional classes.}

\item{value}{Initial value of button, must be an integer.}
}
\description{
A bare button.
}
