% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exp.R
\name{...class}
\alias{...class}
\title{Class}
\usage{
...class(...)
}
\arguments{
\item{...}{Elements, character vectors of any length.}
}
\description{
Build classes from character vectors.
}
\examples{
...class("hello", "world")
...class("hello", c(1, 2))

}
