% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep.R
\name{get_js_dependency}
\alias{get_js_dependency}
\title{JavaScript Dependencies}
\usage{
get_js_dependency(name)
}
\arguments{
\item{name}{Name of file.}
}
\description{
Retrieve JavaScript dependencies.
}
\keyword{internal}
