% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{button}
\alias{button}
\title{Button}
\usage{
button(id, ..., class = ...class(), .tag = htmltools::tags$button)
}
\arguments{
\item{id}{ID of button and \code{input}.}

\item{...}{Content of button.}

\item{class}{Any additional classes.}

\item{.tag}{Tag to use, e.g.: \code{tags$a}.}
}
\description{
A bare button.
}
